增加喇叭控制

+++ Users/GlobalVar.c
@@ -10,10 +10,10 @@ __IO u16 release_delay = 0;
 __IO u8 keyState = 0;
 __IO u8 work_event = 0;
 
-u8 debugData1 = 0;
-u8 debugData2 = 0;
-u8 debugData3 = 0;
-u8 debugData4 = 0;
+u16 debugData1 = 0;
+u16 debugData2 = 0;
+u16 debugData3 = 0;
+u16 debugData4 = 0;
 
 u16 ReadTimeSetting(void)
 {
